(maint) Reformat params inheritane
Prior to this commit multiple classes inherited from the params.pp. This
commit changes the layout so that just the init.pp inherits from the
params and all other classes reference the init for parameters. This
also adds the ability to just call pe_mertics_dashboard instead of
pe_metrics_dashboard::install.