(issue 29) Revert params format to fix spec tests
Prior to this commit, the spec tests were all failing because they could
not find variables from the pe_metrics_dashboard class. This commit
changes the install class to inherit from the params class and the
subclasses to pull default parameters from the install class. This
commit also fixes the broken spec tests that resulted from recent
changes in the module.

This commit resolves #29 and #22