(maint) Update to PDK 1.7.0 and Beaker 4
Prior to this commit the repo was on PDK 1.6.1 and there was no top
limit to the beaker gem. Since Beaker 4 was a breaking change, it needed
to be configured to enable the spec tests moving forward. Otherwise we
would have enounted random failures in the CI unrelated to the changes
in PRs. This commit updates the repo to use Beaker 4.