Refactor to a more standard layout
This commit refactors the module to more closely follow the install >
config > service manifest layout. Parameters were removed on all
manifests except init.pp. All manifests now directly reference
parameters from the puppet_metrics_dashboard class. The greatly reduces
duplication of code and sets the module up in a way that it could easily
move from params.pp to data in module later.

The docs at the top of manifests also got updated to reflect that
puppet-strings expects both a summary and a description that is not
prefixed with the @summary tag. In many cases this means duplicate
info... that is an unfortunate shortcoming of puppet-strings :(

Tests were lightly updated to reflect that subclasses do not have
params. One test was moved form install_spec.rb to init_spec.rb as it
made more sense there.