Stop creating /run/grafana on CentOS 7
This commit updates the service management logic for grafana to drop the
creation of the /run/grafana directory prior to starting the service on
CentOS 7. Grafana 4.5.0 and later manage this by setting the RuntimeDirectory
option in the grafana-server.service unit file. However, CentOS 7 uses systemd
v219 which will fail during startup if the directory specified by
RuntimeDirectory already exists. This failure causes the first catalog
application that installs and starts grafana-server to fail --- which is
particularly problematic for Bolt plans that only run once.

Fixes puppetlabs/puppet_metrics_dashboard#37.