Allow port numbers of services to be specified
This commit updates the master_list, puppetdb_list, and postgres_host_list
parameters to allow pairs of `[hostname, port_number]` in addition to just
a hostname. If the port number is specified, then Telegraf will be configured
to use that port when requesting metrics from the host. If no port number is
specified, then Telegraf will be configured to use the default prts of
8140 for Puppet Server, 8081 for PuppetDB, and 5432 for Postgres.

Closes puppetlabs/puppet_metrics_dashboard#48