Add datasource variable to Archive dashboards
This commit adds a `$datasource` variable to the templates for the
"Archive" dashboards that can be used to quickly flip between different
datasources. This enables metrics from seperate sites to be loaded into
seperate InfluxDB databases and makes them accessible without the need
to duplicate the "Archive" template to create seperate dashboards.