Use shared crosshairs in dashboard templates
This commit changes the value of the `graphTooltip` parameter in each
dashboard template from 0 to 1. Using a value of 1 provides a shared
crosshair that marks the same instant in time the mouse hovers over
on each graph in the dashboard. This enables values to be compared
between graphs that are not stacked on top of each other vertically.