Switching to in module hiera data & updating docs
This commit does the following:

- Switched from Hiera backend to standard function
- Created a function for determining the needed PuppetDB metrics based
  on PE version
- Removed the variable 'storage_metrics_db_queries' since it was not
  referenced anywhere.
- Updated the README and documentation in some manifests

This all stemmed from trying to solve the problem of having to define
the same values for parameter defaults in multiple places. The part of
this commit that addresses that exact problem is the new
puppet_metrics_dashboard::puppetdb_metrics() function. The rest of this
was just a by-product of getting to that solution.