Update .fixtures.yml
This commit updates the test fixtures file to remove pins to explicit
versions of the modules. These pins were out of date with respect to
the version constraints in the metadata.json file and other supported
modules have a pattern of testing against the master branches of their
dependencies instead of pinning to refs. A `puppet_version` constraint
is also added to the `yumrepo_core` module so that it is only used
when testing against Puppet 5 and newer.