(maint) Fix hung spec tests
Prior to this commit, spec tests were hanging indefinately. This commit
fixes that issue by changing the way the facter version was specified in
the spec testing. It sets up a default facter version which is
compatible with the newer versions of `rspec-puppet-facts`. This change
also updates PDK to 1.13.0 to apply the configuration change in the
`.sync.yml`.