(maint) Debian repo tests have multiple codenames
Prior to this commit, the debian based OS spec tests only listed one of
the supported code names for the distro. This caused spec tests to fail
when multiple debian releases were supported. This change enables
checking the repo configuration against the code name in the facts hash
instead of statically.