Use the jruby metrics in the dashboard examples
Prior to this commit, we used the pe-jruby metrics in the example
dashboards. Since these were released into the open source puppet
versions in 5.3.x, we can now use the `jruby` to ensure compatiblity with
OSP and puppetlabs-puppet_metrics_collector/pull/29. This change updates
the dashboards to use `puppetserver.jruby` instead of `puppetserver.pe-jruby`.