Enable the client cert in PuppetDB telegraf metrics
Prior to this commit, telegraf was using insecure SSL connections to
PuppetDB. This worked in some cases, but would fail in other cases due
to not providing a client certificate. As the clent certificate is
available on the host, this commit enables using the client certificate
by default on all PuppetDB queries.