(SUP-2754) WIP for adding SLES support.
This is a fork of #175 with a rebase and fix for an ordering issue with
the systemd::unit_file resource.  Something about this being a defined
type is causing the "No such resource" error, so this commit add an
additional service resource with a require on the unit_file defined
type.