Enable PuppetDB collection on newer PuppetDB versions
Prior to this commit, PuppetDB collection was restricted to localhost on
recent PuppetDB versions by default. It was possible to specify the
PuppetDB nodes manually, but the default parameters would not populate
based on specific PuppetDB versions. This commit updates the logic to
automatically collect data from PuppetDB versions that allow remote
collection.