Enable PuppetDB SSL in dbcompiler profile
This commit updates `puppet_metrics_dashboard::profile::dbcompiler::install`
to pass `enable_client_cert => true` when configuring PuppetDB metrics.
This parameter controls whether `http://` or `https://` is used as the
collection scheme.

As the default port number queried is 8081, which is the SSL port, `https://`
must be used. Additionally, client certificates must be use to authorize
access to the `/status` and `/metrics` API endpoints due to CVE-2020-7943.