Accept 503 responses from PE /status APIs
The puppetlabs/trapperkeeper-status library that provides the
`/status/v1` API of various PE services will respond with a
HTTP 503 status code if any sub-service is not in a healthy state.
By default, Telegraf HTTP inputs will only accept 200 as a
successful response code. This state of affairs leads to situations
like an outage in `pe-orchestration-services` resulting in Telegraf
rejecting perfectly good JRuby metrics from `pe-puppetserver` on
compilers because the `pcp-broker` sub-service is unhealthy.

This commit updates the Telegraf configuration to accept HTTP 503
as a successful return code in addition to 200 so that an attempt
is made to process and store metrics data during outages of external
services.